adding interactive visual test to the relevant file
