+++ src/LinearProgressBar/test/LinearProgressBar.visual.js
@@ -2,6 +2,30 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import LinearProgressBar from '../LinearProgressBar';
 import { storySettings } from './storySettings';
+import { linearProgressBarTestkitFactory } from '../../../testkit';
+
+const { dataHook } = storySettings;
+
+const createLinearProgressBarDriver = dataHook =>
+  linearProgressBarTestkitFactory({
+    wrapper: document.body,
+    dataHook,
+  });
+
+class InteractiveEyeTest extends React.PureComponent {
+  componentDidMount() {
+    this.props.componentDidMount();
+  }
+
+  render() {
+    const props = this.props;
+    return (
+      <div style={{ padding: '40px' }}>
+        <LinearProgressBar dataHook={dataHook} {...props} />
+      </div>
+    );
+  }
+}
 
 const commonProps = {
   dataHook: storySettings.dataHook,
@@ -80,12 +104,41 @@ const tests = [
   },
 ];
 
+const interactiveTests = [
+  {
+    describe: 'Tooltip',
+    its: [
+      {
+        it: 'displayed on icon hover when error progress indication exist',
+        props: {
+          value: 20,
+          showProgressIndication: true,
+          error: true,
+          errorMessage: 'some error message',
+        },
+        componentDidMount: () => {
+          const driver = createLinearProgressBarDriver(dataHook);
+          driver.getTooltip().mouseEnter();
+        },
+      },
+    ],
+  },
+];
+
 tests.forEach(({ describe, its }) => {
   its.forEach(({ it, props }) => {
     storiesOf(`LinearProgressBar/${describe}`, module).add(it, () => (
       <div style={{ width: '40%' }}>
         <LinearProgressBar {...commonProps} {...props} />
       </div>
+    ));
+  });
+});
+
+interactiveTests.forEach(({ describe, its }) => {
+  its.forEach(({ it, props, componentDidMount }) => {
+    storiesOf(`LinearProgressBar/${describe}`, module).add(it, () => (
+      <InteractiveEyeTest {...props} componentDidMount={componentDidMount} />
     ));
   });
 });
