+++ src/LinearProgressBar/test/storySettings.js
@@ -5,7 +5,6 @@ export const storySettings = {
   storyName: 'LinearProgressBar',
   dataHook: 'story-linear-progress-bar',
   testStoryNames: {
-    tooltipOnErrorProgressIndication:
-      'display tooltip on icon hover when there is error progress indication is set to true',
+    DEFAULT: 'default',
   },
 };
